zabbix

+++ sw/spilib/miner_gate/zabbix_log.h
@@ -14,7 +14,7 @@ typedef struct mg_asic_log {
 	uint8_t temp; // ASIC_TEMP
 	uint8_t working_engines;
 	uint8_t failed_bists;
-	uint32_t  freq_time;
+	uint16_t  freq_time;
 } __attribute__((__packed__ )) MG_ASIC_LOG ;
